+++ packages/extension-auth/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @baeta/extension-auth
 
+## 1.0.2
+
+### Patch Changes
+
+- [`1dbd072`](https://github.com/andreisergiu98/baeta/commit/1dbd072c70f8d6faf469b49bd26b6bf45a9bf6ff) Thanks [@andreisergiu98](https://github.com/andreisergiu98)! - fix: return type when returning bool in auth function
+
+- Updated dependencies [[`59bbb9c`](https://github.com/andreisergiu98/baeta/commit/59bbb9c4baaf716f27dc251fe7aeb0231e6c5321), [`d77cd8a`](https://github.com/andreisergiu98/baeta/commit/d77cd8a1810fdf72cfbbb08d05c207bbc893c822), [`cf9f094`](https://github.com/andreisergiu98/baeta/commit/cf9f09468f84d99b069eb0f55e1fc207e2a41dd8)]:
+  - @baeta/core@0.1.2
+  - @baeta/errors@0.1.2
+
 ## 1.0.1
 
 ### Patch Changes